backend/requirements.txt
fastapi==0.110.1
sentry-sdk[fastapi]>=2.0.0
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
SQLAlchemy==2.0.36
alembic==1.13.3
asyncmy==0.2.9

pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas==2.1.4
numpy==1.26.4
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
black==25.9.0
boto3==1.40.39
botocore==1.40.39
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==46.0.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.2
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jq==1.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.77.5
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.6.4
mypy==1.18.2
mypy_extensions==1.1.0
oauthlib==3.3.1
openai>=1.12.0,<2.0.0
anthropic>=0.39.0,<1.0.0
slowapi>=0.1.9
python-json-logger>=3.0.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
pypdf==6.3.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
stripe==12.5.1
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.18
uvicorn==0.25.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
